menu

+++ Restoran/Areas/Admin/Controllers/DishController.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Restoran.Areas.Admin.Data;
 using Restoran.Areas.Admin.Extensions;
@@ -17,21 +18,30 @@ namespace Restoran.Areas.Admin.Controllers
         }
         public async Task<IActionResult> Index()
         {
-          var dishes = await _dbContext.Dishes.ToListAsync();
+          var dishes = await _dbContext.Dishes
+                .Include(d => d.Category)
+                .ToListAsync();
 
           return View(dishes);
         }
         public async Task<IActionResult> Details(int id)
         {
-          var dish = await _dbContext.Dishes.AsNoTracking().SingleOrDefaultAsync(d => d.Id == id);
+          var dish = await _dbContext.Dishes
+                .Include(d => d.Category)
+                .AsNoTracking()
+                .SingleOrDefaultAsync(d => d.Id == id);
 
           return View(dish);
         }
         public async Task<IActionResult> Create()
         {
+            var categories = await _dbContext.Categories.ToListAsync();
+            var categoryListItems = categories.Select(x => new SelectListItem(x.Name, x.Id.ToString())).ToList();
+
             var dishCreateModel = new DishCreateViewModel
             {
                 Title = string.Empty,
+                CategorySelectListItems = categoryListItems,
                 Description = string.Empty,
                 ImageUrl = string.Empty,
                 ImageFile = null
@@ -43,14 +53,21 @@ namespace Restoran.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(DishCreateViewModel model)
         {
+            var categories = await _dbContext.Categories.ToListAsync();
+            var categoryListItems = categories.Select(x => new SelectListItem(x.Name, x.Id.ToString())).ToList();
+
             if (!ModelState.IsValid)
             {
-               return View(model);
+                model.CategorySelectListItems = categoryListItems;
+                ModelState.AddModelError("", "Xeta bas verdi, yeniden yoxlayin!");
+
+                return View(model);
             }
 
             if (!model.ImageFile.IsImage())
             {
                 ModelState.AddModelError("ImageFile", "Sekil secilmelidir!");
+                model.CategorySelectListItems = categoryListItems;
 
                 return View(model);
             }
@@ -58,6 +75,7 @@ namespace Restoran.Areas.Admin.Controllers
             if (!model.ImageFile.IsAllowedSize(1))
             {
                 ModelState.AddModelError("ImageFile", "Sekil hecmi 1mb-dan cox ola bilmez");
+                model.CategorySelectListItems = categoryListItems;
 
                 return View(model);
             }
@@ -69,6 +87,7 @@ namespace Restoran.Areas.Admin.Controllers
                 Title = model.Title,
                 Description = model.Description,
                 Price = model.Price,
+                CategoryId = model.CategoryId,
                 ImageUrl = unicalImageFileName
             };          
 
@@ -94,10 +113,13 @@ namespace Restoran.Areas.Admin.Controllers
 
             return Json(removedDish.Entity);
         }
-        public IActionResult Update(int id)
+        public async Task<IActionResult> Update(int id)
         {
             var dish = _dbContext.Dishes
                 .SingleOrDefault(x => x.Id == id);
+
+            var categories = await _dbContext.Categories.ToListAsync();
+            var categoryListItems = categories.Select(x => new SelectListItem(x.Name, x.Id.ToString())).ToList();
 
             if (dish == null)
                 return NotFound();
@@ -106,6 +128,8 @@ namespace Restoran.Areas.Admin.Controllers
             {
                 Title = dish.Title,
                 Description = dish.Description,
+                CategoryId = dish.CategoryId,
+                CategorySelectListItems = categoryListItems,
                 ImageUrl = dish.ImageUrl,
                 Price = dish.Price,
                 ImageFile = null
@@ -113,6 +137,7 @@ namespace Restoran.Areas.Admin.Controllers
          
             return View(dishUpdateModel);
         }
+
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Update(int id, DishUpdateViewModel model)
@@ -121,35 +146,48 @@ namespace Restoran.Areas.Admin.Controllers
 
             if (dish == null) return NotFound();
 
+            var categories = await _dbContext.Categories.ToListAsync();
+            var categoryListItems = categories.Select(x => new SelectListItem(x.Name, x.Id.ToString())).ToList();
+
             if (!ModelState.IsValid)
             {
+                model.CategorySelectListItems = categoryListItems;
+
                 return View(model);
             }
 
             dish.Title = model.Title;
             dish.Description = model.Description;
             dish.Price = model.Price;
-          
-            if (!model.ImageFile.IsImage())
+            dish.CategoryId = model.CategoryId;
+
+            if (model.ImageFile != null)
             {
-                ModelState.AddModelError("ImageFile", "Sekil secilmelidir!");
-                return View(model);
-            }
+                if (!model.ImageFile.IsImage())
+                {
+                    ModelState.AddModelError("ImageFile", "Sekil secilmelidir!");
+                    model.CategorySelectListItems = categoryListItems;
 
-            if (!model.ImageFile.IsAllowedSize(1))
-            {
-                ModelState.AddModelError("ImageFile", "Sekil hecmi 1mb-dan cox ola bilmez");
-                return View(model);
-            }
+                    return View(model);
+                }
 
-            var unicalImageFileName = await model.ImageFile.GenerateFile(FilePathConstants.DishPath);
+                if (!model.ImageFile.IsAllowedSize(1))
+                {
+                    ModelState.AddModelError("ImageFile", "Sekil hecmi 1mb-dan cox ola bilmez");
+                    model.CategorySelectListItems = categoryListItems;
 
-            if (dish.ImageUrl != null)
-            {
-                System.IO.File.Delete(Path.Combine(FilePathConstants.DishPath, dish.ImageUrl));
-            }
+                    return View(model);
+                }
 
-            dish.ImageUrl = unicalImageFileName;
+                var unicalImageFileName = await model.ImageFile.GenerateFile(FilePathConstants.DishPath);
+
+                if (dish.ImageUrl != null)
+                {
+                    System.IO.File.Delete(Path.Combine(FilePathConstants.DishPath, dish.ImageUrl));
+                }
+
+                 dish.ImageUrl = unicalImageFileName;
+            }    
 
             _dbContext.Dishes.Update(dish);
             await _dbContext.SaveChangesAsync();
